tests/wpt/web-platform-tests/2dcontext/fill-and-stroke-styles/2d.gradient.object.type.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.gradient.object.type</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.gradient.object.type</h1>
<p class="desc">window.CanvasGradient exists and has the right properties</p>

<p class="notes">Defined in "Web IDL" (draft)
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("window.CanvasGradient exists and has the right properties");
_addTest(function(canvas, ctx) {

_assertDifferent(window.CanvasGradient, undefined, "window.CanvasGradient", "undefined");
_assertDifferent(window.CanvasGradient.prototype.addColorStop, undefined, "window.CanvasGradient.prototype.addColorStop", "undefined");


});
</script>
